icecap-qemu: Replaced process stdio inheritance with io::copy loop
This is necessary due to locking issues under Cargo test. It seems that
even in --nocapture mode, the stdout of the process is replaced by a
sort of synchronized stdout. The problem is when we inherit the stdout
in a sub-process, we end up corrupting this synchronization mechanism,
cause the tests to _sometimes_ fail with "failed printing to stdout:
Resource temporarily unavailable".

This is an inferior solution, in that it creates two otherwise
unnecessary threads for a problem in the test framework, but it
works and should be more resilient to other changes in the future.